Handle a case in highlight logic where the exact amount of characters is known but no correct characters are known.
If the user manages to find the count of character in word before finding the correct index for it,
for instance if word is ABBBB, guess is BBAAA => only one A is highlighted, count of A is known as 1, but no correct
index is known yet.

Well, user could probably often already deduce the correct indices, but they haven't been guessed yet so highlight logic
won't do that for the end user either.